backend/pyproject.toml
[tool.poetry]
name = "challenge-fastapi"
version = "0.1.0"
description = ""
authors = ["Pei <perezcastagnino.andres@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
psycopg2 = "^2.9.10"
sqlmodel = "^0.0.22"
bcrypt = "^4.2.0"
passlib = "^1.7.4"
python-multipart = "^0.0.16"
python-jose = "^3.3.0"
apscheduler = "^3.10.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.13.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
